fix: entities type and reward
