+++ apps/collector-page/src/components/business/task/ExploreTaskSwiperItem.tsx
@@ -2,14 +2,14 @@
  * @Author: shixuewen friendlysxw@163.com
  * @Date: 2022-07-05 18:35:10
  * @LastEditors: shixuewen friendlysxw@163.com
- * @LastEditTime: 2022-08-09 16:35:59
+ * @LastEditTime: 2022-08-15 14:49:42
  * @Description: file description
  */
 import React from 'react'
 import { useNavigate } from 'react-router-dom'
 import styled from 'styled-components'
 import { ScrollBarCss } from '../../../GlobalStyle'
-import { TaskType } from '../../../types/api'
+import { RewardType } from '../../../types/entities'
 import IconAlarmClock from '../../common/icons/IconAlarmClock'
 import IconGiftBox from '../../common/icons/IconGiftBox'
 import OverflowEllipsisBox from '../../common/text/OverflowEllipsisBox'
@@ -20,7 +20,6 @@ export type ExploreTaskSwiperItemDataType = {
   id: number
   name: string
   image: string
-  type: TaskType
   startTime: number
   endTime: number
   winnerNum: number
@@ -32,6 +31,8 @@ export type ExploreTaskSwiperItemDataType = {
   }
   reward?: {
     name: string
+    type: RewardType
+    raffled: boolean
   }
 }
 export type ExploreTaskSwiperItemViewConfigType = {}
@@ -44,18 +45,23 @@ export type ExploreTaskSwiperItemHandlesType = {}
 
 export type ExploreTaskSwiperItemProps = ExploreTaskSwiperItemDataViewType & ExploreTaskSwiperItemHandlesType
 
-const TaskTypeLabels = {
-  [TaskType.WHITELIST_ORIENTED]: 'Whitelist-Oriented Task',
-  [TaskType.WHITELIST_LUCK_DRAW]: 'Whitelist Luck Draw',
-}
-
 const ExploreTaskSwiperItem: React.FC<ExploreTaskSwiperItemProps> = ({
   data,
   viewConfig,
 }: ExploreTaskSwiperItemProps) => {
   const navigate = useNavigate()
-  const { id, name, image, type, startTime, endTime, winnerNum, description, project, reward } = data
-  const taskTypeLabel = TaskTypeLabels[type] || 'Unknown Task Type'
+  const { id, name, image, startTime, endTime, winnerNum, description, project, reward } = data
+  let taskTypeLabel = 'Unknown Reward Type'
+  if (reward) {
+    switch (reward.type) {
+      case RewardType.WHITELIST:
+        taskTypeLabel = reward.raffled ? 'Whitelist Luck Draw' : 'Whitelist-Oriented Task'
+        break
+      case RewardType.OTHERS:
+        taskTypeLabel = 'Other Reward'
+        break
+    }
+  }
   const startDate = new Date(startTime).toLocaleDateString()
   const endDate = new Date(endTime).toLocaleDateString()
   return (
